+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.1 (WIP)
 
 * Move routes to separate `routes.js` file, so they load in the correct order
+* QR-codes will now link to `/qr/{shortcode}`, so we can track QR-code usages
 
 ## 0.2.0 (2023-01-24)